requirements.txt
tensorflow
numpy
pandas
python-binance
scikit-learn
Pillow
keras